memory/handlers/start_handler.py
# memory/handlers/start_handler.py

from telegram import Update
from telegram.ext import ContextTypes
import logging

logger = logging.getLogger(__name__)

async def start_handler(update: Update, context: ContextTypes.DEFAULT_TYPE):
    logger.debug(f"Received /start command from user {update.effective_user.id}")
    await update.message.reply_text("Welcome! Send me a tweet link to store it.")