requirements.txt
python-telegram-bot==20.3
boto3==1.28.59
python-dotenv==1.0.0